Refactor project structure, update design, and enhance BNG Shop features
